Refactored ConsoleManager
